Updated MSpec source to d2e7a635.

+++ mspec/lib/mspec/helpers/tmp.rb
@@ -7,9 +7,13 @@ SPEC_TEMP_DIR = "#{File.expand_path(Dir.pwd)}/rubyspec_temp"
 
 SPEC_TEMP_UNIQUIFIER = "0"
 
+SPEC_TMEM_DIR_PID = Process.pid
+
 at_exit do
   begin
-    Dir.delete SPEC_TEMP_DIR if File.directory? SPEC_TEMP_DIR
+    if SPEC_TMEM_DIR_PID == Process.pid
+      Dir.delete SPEC_TEMP_DIR if File.directory? SPEC_TEMP_DIR
+    end
   rescue SystemCallError
     STDERR.puts <<-EOM
